// src/app/shared/types/apis.ts
export interface IApis {
  liveTvApi: any;
  allSportApi: string;
  sslSportApi: any;
  sslAllSportSocketApi: any;
  sslAllSportApi:any;
  premiumApi: string;
  ip: string;
  adminIp: string;
  casinoApi: string;
  exchangeGamesApi: string;
  fancyApi: string;
  racingApi: string;
  sportApi: string;
  virtualSportApi: string;
  sportSocketApi: string;
  racingSocketApi: string;
  scoreApi:string;
}